+++ commands/__init__.py
@@ -3,5 +3,6 @@ from commands.album_commands import Album
 from commands.scout_commands import Scout
 from commands.stats_commands import Stats
 from commands.trivia_commands import Trivia
+from commands.config_commands import Config
 
-__all__ = ['Scout', 'Album', 'Info', 'Stats', 'Trivia']
+__all__ = ['Scout', 'Album', 'Info', 'Stats', 'Trivia', 'Config']