+++ main.py
@@ -34,7 +34,14 @@ def main():
     )
 
     bot.remove_command('help')
-    cogs = [Scout(bot), Album(bot), Info(bot), Stats(bot), Trivia(bot)]
+    cogs = [
+        Scout(bot), 
+        Album(bot), 
+        Info(bot), 
+        Stats(bot), 
+        Trivia(bot), 
+        Config(bot)
+    ]
 
     card_update_thread = Thread(target=update_task)
     card_update_thread.setDaemon(True)